requirements.txt
qdrant-client
numpy
python-dotenv
